refactor: 모달 내 명함 숨김/삭제 동작 코드 정리 #19
